+++ NEWS.md
@@ -57,6 +57,8 @@
 
 * Fixed issues with contrasting slopes when `backend` was `"emmeans"`.
 
+* Fixed issues in `estimate_contrasts()` when filtering numeric values in `by`.
+
 * Fixed issue in `estimate_slopes()` for models from package *lme4*.
 
 # modelbased 0.9.0